Make the layer direction a rule, and move the code that broke it
`services/giveawayService` required `commands/utility/giveaway` for
`endGiveaway`. `services/scheduler` required a command to pick up unclosed polls
at startup, and `services/ai/actions` required one to draw a poll embed.
`utils/logger` required `services/caseService`; `utils/applyXpGain` required
`services/petService`. Every one is a lower layer reaching up into a higher one,
which is the direction a require cycle arrives from — and in CommonJS a cycle is
not a warning, it is a module object that is half-empty at require time. The
heist button handler had already met that: it required its own service lazily,
from inside the function, to dodge the cycle its position created.

None of the shared code was command-specific, so it moves down rather than being
re-exported sideways:

  views/pollView, views/heistView   the embeds and component rows several
                                    callers draw. A new layer, below services,
                                    which may import nothing but discord.js and
                                    static data.
  services/pollService              persistence, voting and expiry. The command
                                    keeps its definition and its reply.
  services/giveawayService          the draw and the ending. `/giveaway end` and
                                    the scheduled sweep are two callers of one
                                    operation.
  services/heistService             skill checks, resolution, the lobby
                                    countdown and the button handler. The lazy
                                    require is gone with the cycle.
  data/heistData                    ROLES and TARGETS, so the view can lay out a
                                    lobby without importing a service.
  services/moderationLogService     was utils/logger. It reads a Guild, writes a
                                    Case and talks to Discord; it was never a
                                    utility.
  services/levelingService          was utils/applyXpGain. Applying XP *is*
                                    consulting the pet passive, so it belongs
                                    with the service it consults.

Then the direction is enforced: `layers/no-upward-require`, a local ESLint rule,
fails the build on a module requiring one from a layer above it. It is a rule
rather than a test because a lint error lands in the editor on the line that
caused it, and CI already runs `npm run lint`. It is written here rather than
pulled in as eslint-plugin-boundaries because core ESLint dropped
`no-restricted-modules` and nothing in core matches a `require()` path — this is
forty lines, and it is the whole of what was needed.

One exception is listed, with its reason: Mongoose only honours middleware
registered before model() compiles the schema, so the guild-settings cache
invalidation has to be attached in models/Guild.js. That require is already
lazy for the same reason.

Closes #614.
