+++ src/commands/utility/poll.js
@@ -1,6 +1,6 @@
-const { SlashCommandBuilder, EmbedBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle, MessageFlags } = require('discord.js');
-const Poll = require('../../models/Poll');
-const { runJob } = require('../../utils/jobRunner');
+const { SlashCommandBuilder, MessageFlags } = require('discord.js');
+const { parseDuration, createPoll } = require('../../services/pollService');
+const { buildPollEmbed, buildPollRows } = require('../../views/pollView');
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -35,174 +35,14 @@ module.exports = {
         await interaction.reply({ embeds: [embed], components: rows });
         const msg = await interaction.fetchReply();
 
-        await Poll.create({
-            messageId: msg.id,
+        await createPoll({
+            msg,
             guildId: interaction.guild.id,
             channelId: interaction.channel.id,
             question,
             options,
-            votes: new Map(),
             endsAt,
-            createdBy: interaction.user.tag
+            createdBy: interaction.user.tag,
         });
-
-        if (endsAt) {
-            scheduleExpiry(msg, question, options, endsAt, interaction.user.tag);
-        }
     }
 };
-
-function parseDuration(str) {
-    const match = str.match(/^(\d+)(s|m|h|d)$/i);
-    if (!match) return null;
-    const multipliers = { s: 1000, m: 60000, h: 3600000, d: 86400000 };
-    return parseInt(match[1]) * multipliers[match[2].toLowerCase()];
-}
-
-function tallyVotes(voteMap, optionCount) {
-    const counts = new Array(optionCount).fill(0);
-    for (const idx of voteMap.values()) counts[idx]++;
-    return counts;
-}
-
-function buildPollEmbed(question, options, counts, endsAt, author, closed = false) {
-    const total = counts.reduce((a, b) => a + b, 0);
-
-    const lines = options.map((opt, i) => {
-        const pct = total > 0 ? Math.round((counts[i] / total) * 100) : 0;
-        const bar = '█'.repeat(Math.round(pct / 10)) + '░'.repeat(10 - Math.round(pct / 10));
-        return `**${i + 1}. ${opt}**\n${bar} ${pct}% (${counts[i]} vote${counts[i] !== 1 ? 's' : ''})`;
-    });
-
-    const authorTag = typeof author === 'string' ? author : author.tag;
-
-    const embed = new EmbedBuilder()
-        .setColor(closed ? '#ff0000' : '#5865F2')
-        .setTitle(`${closed ? '🔒 ' : '📊 '}${question}`)
-        .setDescription(lines.join('\n\n'))
-        .addFields({ name: 'Total votes', value: total.toString(), inline: true })
-        .setFooter({ text: `Created by ${authorTag}${closed ? ' • Poll closed' : ''}` });
-
-    if (endsAt && !closed) {
-        embed.addFields({ name: 'Ends', value: `<t:${Math.floor(endsAt.getTime() / 1000)}:R>`, inline: true });
-    }
-
-    return embed;
-}
-
-function buildPollRows(options) {
-    const rows = [];
-    for (let i = 0; i < options.length; i += 5) {
-        const row = new ActionRowBuilder();
-        options.slice(i, i + 5).forEach((opt, j) => {
-            row.addComponents(
-                new ButtonBuilder()
-                    .setCustomId(`poll_${i + j}`)
-                    .setLabel(`${i + j + 1}. ${opt.substring(0, 77)}`)
-                    .setStyle(ButtonStyle.Secondary)
-            );
-        });
-        rows.push(row);
-    }
-    return rows;
-}
-
-// Discord API error codes meaning the poll message can never be edited again:
-// Unknown Message, Unknown Channel, Missing Access.
-const MESSAGE_UNREACHABLE = new Set([10008, 10003, 50001]);
-
-function scheduleExpiry(msg, question, options, endsAt, createdBy) {
-    // Overdue polls close now rather than being dropped. A poll only reaches
-    // this already past its end when it is being picked up at startup — the bot
-    // was down when it expired, or a transient edit failure left it open — and
-    // returning early there is what made "still open" permanent.
-    const delay = Math.max(0, endsAt.getTime() - Date.now());
-    // Through runJob like every other scheduled callback (#611): a poll that
-    // fails to close is a poll that stays open forever, and a swallowed
-    // console.error was the only trace of it. `scope` is the message id
-    // because this fires once for one poll and no later tick retries it — two
-    // polls expiring in the same second must both run, not one be dropped as
-    // an overlap of the other.
-    setTimeout(() => runJob('poll', 'closeExpiredPoll', async () => {
-        const poll = await Poll.findOne({ messageId: msg.id });
-        if (!poll || poll.closed) return;
-
-        const counts = tallyVotes(poll.votes, options.length);
-        const closedEmbed = buildPollEmbed(question, options, counts, endsAt, createdBy, true);
-        try {
-            await msg.edit({ embeds: [closedEmbed], components: [] });
-        } catch (err) {
-            // A message that is not there any more is the one failure that
-            // still ends the poll: there is nothing left to edit and nothing a
-            // retry would reach, so mark it closed and stop taking votes.
-            // Anything else — a rate limit, an outage — leaves the poll open
-            // with live buttons, which is the truth, and propagates so runJob
-            // records it instead of the poll being marked closed against a
-            // message still showing it as running.
-            if (!MESSAGE_UNREACHABLE.has(err?.code)) throw err;
-        }
-
-        poll.closed = true;
-        await poll.save();
-    }, { guildId: msg.guildId, scope: msg.id, payload: { messageId: msg.id } }), delay);
-}
-
-async function handlePollVote(interaction) {
-    const optionIndex = parseInt(interaction.customId.split('_')[1]);
-    const messageId = interaction.message.id;
-
-    const poll = await Poll.findOne({ messageId });
-    if (!poll) {
-        return interaction.reply({ content: 'This poll is no longer active.', flags: MessageFlags.Ephemeral });
-    }
-    if (poll.closed) {
-        return interaction.reply({ content: 'This poll is closed.', flags: MessageFlags.Ephemeral });
-    }
-
-    const existing = poll.votes.get(interaction.user.id);
-    if (existing === optionIndex) {
-        poll.votes.delete(interaction.user.id);
-        await interaction.reply({ content: 'Your vote has been removed.', flags: MessageFlags.Ephemeral });
-    } else {
-        poll.votes.set(interaction.user.id, optionIndex);
-        await interaction.reply({ content: `You voted for option **${optionIndex + 1}**.`, flags: MessageFlags.Ephemeral });
-    }
-    poll.markModified('votes');
-    await poll.save();
-
-    const counts = tallyVotes(poll.votes, poll.options.length);
-    const newEmbed = buildPollEmbed(poll.question, poll.options, counts, poll.endsAt, poll.createdBy);
-    const rows = buildPollRows(poll.options);
-    await interaction.message.edit({ embeds: [newEmbed], components: rows }).catch(() => {});
-}
-
-async function scheduleActivePollExpirations(client) {
-    try {
-        // No lower bound on endsAt: a poll that expired while the process was
-        // down has no timer left anywhere, so excluding it here is what left it
-        // open forever. scheduleExpiry closes an overdue one immediately.
-        const active = await Poll.find({ closed: false });
-        for (const poll of active) {
-            try {
-                const guild = client.guilds.cache.get(poll.guildId);
-                if (!guild) continue;
-                const channel = guild.channels.cache.get(poll.channelId);
-                if (!channel) continue;
-                const msg = await channel.messages.fetch(poll.messageId).catch(() => null);
-                if (!msg) continue;
-                scheduleExpiry(msg, poll.question, poll.options, poll.endsAt, poll.createdBy);
-            } catch (err) {
-                console.error('[poll] failed to reschedule poll', poll.messageId, err);
-            }
-        }
-        if (active.length) console.log(`[POLL] Picked up ${active.length} unclosed poll(s).`);
-    } catch (err) {
-        console.error('[poll] scheduleActivePollExpirations error:', err);
-    }
-}
-
-module.exports.handlePollVote = handlePollVote;
-module.exports.buildPollEmbed = buildPollEmbed;
-module.exports.buildPollRows = buildPollRows;
-module.exports.tallyVotes = tallyVotes;
-module.exports.scheduleActivePollExpirations = scheduleActivePollExpirations;
