+++ src/events/messageCreate.js
@@ -3,7 +3,7 @@ const Guild = require('../models/Guild');
 const Case = require('../models/Case');
 const Reminder = require('../models/Reminder');
 const { handleAIChat } = require('../services/aiService');
-const { logModeration } = require('../utils/logger');
+const { logModeration } = require('../services/moderationLogService');
 const { ensureQuests, onMessage, onStreakUpdate, notifyQuestComplete, notifyQuestNearComplete, notifyDailyQuestReset } = require('../services/questService');
 const { getStreakMultiplier, checkNewMilestones } = require('../utils/streakMultiplier');
 const { hasEffect, consumeEffect, getXpMultiplier, getServerXpMultiplier } = require('../services/effectsService');
@@ -11,7 +11,7 @@ const { checkRivalry } = require('../services/rivalryService');
 const { checkAndAward, announceAchievements } = require('../services/achievementService');
 const { checkAndBroadcastWealthMilestone } = require('../utils/wealthMilestone');
 const { maybeTriggerChatEvent } = require('../services/chatEventService');
-const { applyXpGain, announceLevelUp } = require('../utils/applyXpGain');
+const { applyXpGain, announceLevelUp } = require('../services/levelingService');
 const BASE_BAD_WORDS = require('../data/profanityList');
 const { getGuildSettings } = require('../utils/guildSettingsCache');
 const { saveWithBalanceDelta } = require('../utils/balanceDelta');
