+++ tests/schedulerOwnsJobs.test.js
@@ -48,7 +48,7 @@ const NON_JOB_TIMERS = new Set([
     'utils/imageRateLimit.js',          // in-memory rate-limit bucket sweep
     'games/casino/crash.js',            // per-lobby round tick
     'commands/fun/meme.js',             // in-memory meme cache sweep
-    'commands/economy/heist.js',        // per-heist lobby countdown
+    'services/heistService.js',         // per-heist lobby countdown
     'commands/economy/syndicate.js',    // per-raid lobby countdown
     'commands/economy/quiz.js',         // per-question countdown
 ]);
@@ -183,7 +183,7 @@ describe('the scheduler owns recurring work', () => {
             summaryService: 'services/summaryService.js',
             'caseService.slaMonitor': 'services/caseService.js',
             questService: 'services/questService.js',
-            'poll.expirations': 'commands/utility/poll.js',
+            'poll.expirations': 'services/pollService.js',
             dailyBibleService: 'services/dailyBibleService.js',
             'rssService.dailyNews': 'services/rssService.js',
         };
